refactor(knockdog): service tag enum 정리
